frontend/requirements.txt
altair==5.0.1
attrs==23.1.0
blinker==1.6.2
branca==0.6.0
cachetools==5.3.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cycler==0.10.0
decorator==5.1.1
folium==0.14.0
fonttools==4.40.0
gitdb==4.0.10
GitPython==3.1.31
idna==2.10
importlib-metadata==6.7.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
numpy==1.25.0
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.23.3
pyarrow==12.0.1
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==2.4.7
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.2
roboflow==1.1.0
six==1.16.0
smmap==5.0.0
streamlit==1.23.1
supervision==0.11.1
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
wget==3.2
zipp==3.15.0
